[Add] Apollo Setup + [Fix] theme state
