+++ package.json
@@ -8,6 +8,7 @@
         "eject": "expo eject"
     },
     "dependencies": {
+        "@apollo/client": "^3.3.6",
         "@expo/vector-icons": "^12.0.2",
         "@react-native-community/masked-view": "0.1.10",
         "@react-navigation/bottom-tabs": "^5.11.2",
@@ -18,6 +19,7 @@
         "@types/react-redux": "^7.1.14",
         "expo": "~40.0.0",
         "expo-status-bar": "~1.0.3",
+        "graphql": "^15.4.0",
         "react": "16.13.1",
         "react-dom": "16.13.1",
         "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
