+++ utils/apolloClient.ts
@@ -1,0 +1,14 @@
+import { ApolloClient, InMemoryCache } from '@apollo/client';
+import { GRAPHQL_API } from './constants';
+
+const client = new ApolloClient({
+    uri: GRAPHQL_API,
+    cache: new InMemoryCache(),
+    defaultOptions: {
+        watchQuery: {
+            fetchPolicy: 'cache-and-network',
+        },
+    },
+});
+
+export default client;
