+++ utils/constants.ts
@@ -1,0 +1,1 @@
+export const GRAPHQL_API = 'http://localhost:8000/graphql';
